updated cart.js with shoppingCart.mjs to working with template
