+++ src/checkout/index.html
@@ -15,7 +15,7 @@
   </head>
 
   <body>
-    <header class="divider">
+    <header class="divider" id="main-header">
       <!-- <div class="logo">
         <img src="images/noun_Tent_2517.svg" alt="tent image for logo" />
 
@@ -65,6 +65,6 @@
       </section>
     </main>
 
-    <footer><!-- &copy;NOT a real business --></footer>
+    <footer id="main-footer"><!-- &copy;NOT a real business --></footer>
   </body>
 </html>
