+++ src/index.html
@@ -114,6 +114,6 @@
         </ul>
       </section>
     </main>
-    <footer><!-- &copy;NOT a real business --></footer>
+    <footer id="main-footer"><!-- &copy;NOT a real business --></footer>
   </body>
 </html>
