+++ src/js/shoppingCart.mjs
@@ -1,0 +1,33 @@
+import { getLocalStorage, renderWithTemplate } from "./utils.mjs";
+/*
+export function renderCartContents() {
+    const cartItems = getLocalStorage('so-cart');
+    const htmlItems = cartItems.map((item) => cartItemTemplate(item));
+    document.querySelector('.product-list').innerHTML = htmlItems.join('');
+}*/
+
+export function shoppingCart() {
+    const cartItems = getLocalStorage('so-cart');
+    console.log("cart Items: ", cartItems);
+    const outputEl  = document.querySelector('.product-list');
+    renderWithTemplate(cartItemTemplate, outputEl, cartItems);
+}
+
+function cartItemTemplate(item) {
+    const newItem = `<li class="cart-card divider">
+    <a href="#" class="cart-card__image">
+      <img
+        src="${item.Image}"
+        alt="${item.Name}"
+      />
+    </a>
+    <a href="#">
+      <h2 class="card__name">${item.Name}</h2>
+    </a>
+    <p class="cart-card__color">${item.Colors[0].ColorName}</p>
+    <p class="cart-card__quantity">qty: 1</p>
+    <p class="cart-card__price">$${item.FinalPrice}</p>
+  </li>`;
+
+    return newItem;
+}
