+++ src/js/utils.mjs
@@ -85,8 +85,8 @@ function loadTemplate(path) {
 export async function loadHeaderFooter() {
   const headerTemplate = loadTemplate("/partials/header.html");
   const footerTemplate = loadTemplate("/partials/footer.html");
-  const headerEl = document.querySelector("header");
-  const footerEl = document.querySelector("footer");
+  const headerEl = document.querySelector("#main-header");
+  const footerEl = document.querySelector("#main-footer");
   renderWithTemplate(headerTemplate, headerEl);
   renderWithTemplate(footerTemplate, footerEl);
 }
