+++ src/product_pages/index.html
@@ -14,7 +14,7 @@
   </head>
 
   <body>
-    <header class="divider">
+    <header class="divider" id="main-header">
      <!-- <div class="logo">
         <img src="/images/noun_Tent_2517.svg" alt="tent image for logo" />
         <a href="../index.html"> Sleep<span class="highlight">Outside</span></a>
@@ -68,6 +68,6 @@
       </section>
     </main>
 
-    <footer><!-- &copy;NOT a real business --></footer>
+    <footer id="main-footer"><!-- &copy;NOT a real business --></footer>
   </body>
 </html>
